+++ src/models/Settings.php
@@ -10,6 +10,7 @@ use craft\base\ElementInterface;
 use craft\base\Model;
 use craft\db\Table;
 use craft\elements\Asset;
+use craft\helpers\App;
 use craft\helpers\ArrayHelper;
 use craft\helpers\Db;
 use craft\helpers\Json;
@@ -187,5 +188,15 @@ class Settings extends Model
         $notificationAdmins = $this->notificationAdmins ?: [];
 
         return ArrayHelper::where($notificationAdmins, 'enabled');
+    }
+
+    public function getRecaptchaKey()
+    {
+        return App::parseEnv($this->recaptchaKey);
+    }
+
+    public function getRecaptchaSecret()
+    {
+        return App::parseEnv($this->recaptchaSecret);
     }
 }